Moved views to different pages and set map as main page
